Documentation for bucket.jl

SoilDynamics._calc_bucket_pos! — Method

_calc_bucket_pos!(
    out::SimOut{B,I,T}, pos::Vector{T}, ori::Quaternion{T}, grid::GridParam{I,T},
    bucket::BucketParam{I,T}, step_bucket_grid::T=0.5, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function determines all the cells where the bucket is located. The bucket position is calculated based on its reference pose stored in the bucket struct, as well as the provided position (pos) and orientation (ori). pos and ori are used to apply the appropriate translation and rotation to the bucket relative to its reference pose. The center of rotation is assumed to be the bucket origin. The orientation is provided using the quaternion definition.

Note

This function is intended for internal use only.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

pos::Vector{Float64}: Cartesian coordinates of the bucket origin. [m]

ori::Quaternion{Float64}: Orientation of the bucket. [Quaternion]

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

bucket::BucketParam{Float64}: Struct that stores information related to the bucket object.

step_bucket_grid::Float64: Spatial increment used to decompose the edges of the bucket.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

pos = [0.5, 0.3, 0.4]
ori = angle_to_quat(0.0, -pi / 2, 0.0, :ZYX)
grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
o = [0.0, 0.0, 0.0]
j = [0.0, 0.0, 0.0]
b = [0.0, 0.0, -0.5]
t = [1.0, 0.0, -0.5]
bucket = BucketParam(o, j, b, t, 0.5)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_calc_bucket_pos!(out, pos, ori, grid, bucket)

SoilDynamics._calc_line_pos — Method

_calc_line_pos(
    a::Vector{T}, b::Vector{T}, delta::T, grid::GridParam{I,T}
) where {I<:Int64,T<:Float64}

This function determines all the cells that lie on a straight line between two Cartesian coordinates.

For the sake of accuracy, the line is divided into smaller segments using a spatial increment delta.

The coordinates of each sub-point (ab_i) along the line can then be calculated as

ab_i = a + ab * i * delta / norm(ab)

where i is the increment number and ab = b - a. The Cartesian coordinates can then be converted into indices

ab_i_ind = ab_i / cell_size + grid_half_length + 1

Finally, the floating-point values are rounded to obtain the cell indices in the X, Y, Z directions. As the center of each cell is considered to be on the center of the top surface, round should be used for getting the cell indices in the X and Y direction, while ceil should be used for the Z direction.

Note

This function is intended for internal use only.

When the line follows a cell border, the exact location of the line becomes ambiguous. It is assumed that the caller resolves this ambiguity.

Inputs

a::Vector{Float64}: Cartesian coordinates of the first extremity of the line. [m]

b::Vector{Float64}: Cartesian coordinates of the second extremity of the line. [m]

delta::Float64: Spatial increment used to decompose the line. [m]

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

Outputs

line_pos::Vector{Vector{Int64}}: Collection of cells indices where the line is located. Result is not sorted and duplicates should be expected.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
a = [1.0, 0.5, 0.7]
b = [0.7, 0.8, -0.3]

line_pos = _calc_line_pos(a, b, 0.01, grid)

SoilDynamics._calc_rectangle_pos — Method

_calc_rectangle_pos(
    a::Vector{T}, b::Vector{T}, c::Vector{T}, d::Vector{T},
    delta::T, grid::GridParam{I,T}, tol::T=1e-8
) where {I<:Int64,T<:Float64}

This function determines the cells where a rectangle surface is located. The rectangle is defined by providing the Cartesian coordinates of its four vertices in the proper order.

To optimize performance, the function iterates over a portion of the horizontal grid where the rectangle is located. For each cell, the function calculates the height of the plane formed by the rectangle at the top right corner of the cell. If the cell is within the rectangle area, the calcualted height is added to the results for the four neighboring cells.

This method works because when a plane intersects with a rectangular cell, the minimum and maximum height of the plane within the cell occurs at one of the cell corners. By iterating through all the cells, the function ensures that all the corners of each cell are investigated.

However, this approach does not work when the rectangle is perpendicular to the XY plane. To handle this case, the function uses the _calc_line_pos function to include the cells that lie on the four edges of the rectangle.

Note

This function is intended for internal use only.

The iteration is performed over the top right corner of each cell, but any other corner could have been chosen without affecting the results.

Not all cells are provided, since, at a given XY position, only the cells with the minimum and maximum height are important.

When the rectangle follows a cell border, the exact location of the rectangle becomes ambiguous. It is assumed that the caller resolves this ambiguity.

Inputs

a::Vector{Float64}: Cartesian coordinates of one vertex of the rectangle. [m]

b::Vector{Float64}: Cartesian coordinates of one vertex of the rectangle. [m]

c::Vector{Float64}: Cartesian coordinates of one vertex of the rectangle. [m]

d::Vector{Float64}: Cartesian coordinates of one vertex of the rectangle. [m]

delta::Float64: Spatial increment used to decompose the edges of the rectangle. [m]

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

Vector{Vector{Int64}}: Collection of cells indices where the rectangle is located. Result is not sorted and duplicates may be present.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
a = [1.0, 0.0, 0.7]
b = [0.0, 1.0, 0.7]
c = [0.0, 1.0, 0.9]
d = [1.0, 0.0, 0.9]

rect_pos = _calc_rectangle_pos(a, b, c, d, 0.01, grid)

SoilDynamics._calc_triangle_pos — Method

_calc_triangle_pos(
    a::Vector{T}, b::Vector{T}, c::Vector{T},
    delta::T, grid::GridParam{I,T}, tol::T=1e-8
) where {I<:Int64,T<:Float64}

This function determines the cells where a triangle surface is located. The triangle is defined by providing the Cartesian coordinates of its three vertices in the proper order.

To optimize performance, the function iterates over a portion of the horizontal grid where the triangle is located. For each cell, the function calculates the height of the plane formed by the triangle at the top right corner of the cell. If the cell is within the triangle area, the calcualted height is added to the results for the four neighboring cells.

This method works because when a plane intersects with a rectangular cell, the minimum and maximum height of the plane within the cell occurs at one of the cell corners. By iterating through all the cells, the function ensures that all the corners of each cell are investigated.

However, this approach does not work when the triangle is perpendicular to the XY plane. To handle this case, the function uses the _calc_line_pos function to include the cells that lie on the three edges of the triangle.

Note

This function is intended for internal use only.

The iteration is performed over the top right corner of each cell, but any other corner could have been chosen without affecting the results.

Not all cells are provided, since, at a given XY position, only the cells with the minimum and maximum height are important.

When the triangle follows a cell border, the exact location of the triangle becomes ambiguous. It is assumed that the caller resolves this ambiguity.

Inputs

a::Vector{Float64}: Cartesian coordinates of one vertex of the triangle. [m]

b::Vector{Float64}: Cartesian coordinates of one vertex of the triangle. [m]

c::Vector{Float64}: Cartesian coordinates of one vertex of the triangle. [m]

delta::Float64: Spatial increment used to decompose the edges of the triangle. [m]

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

Vector{Vector{Int64}}: Collection of cells indices where the triangle is located. Result is not sorted and duplicates may be present.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
a = [1.0, 0.0, 0.7]
b = [0.0, 1.0, 0.7]
c = [0.0, 1.0, 0.9]

tri_pos = _calc_triangle_pos(a, b, c, 0.01, grid)

SoilDynamics._decompose_vector_rectangle — Method

_decompose_vector_rectangle(
    ab_ind::Vector{T}, ad_ind::Vector{T}, a_ind::Vector{T},
    area_min_x::I, area_min_y::I, area_length_x::I, area_length_y::I, tol::T=1e-8
) where {I<:Int64,T<:Float64}

This function performs a vector decomposition on a portion of the horizontal plane where a rectangle ABCD is located. The position of the rectangle is defined by its edges AB and AD, while the specified area extends over [area_min_x, area_min_x + area_length_x] on the X direction and [area_min_y, area_min_y + area_length_y] on the Y direction.

For each cell in the specified area, the function decomposes it into the basis formed by the vectors AB and AD. Let O be the name of a cell, it can then be decomposed as

AO = c_ab * AB + c_ad * AD.

This decomposition leads to a system of 2 equations with 2 unknowns (cab and cad)

AO[1] = c_ab * AB[1] + c_ad * AD[1] {1},
AO[2] = c_ab * AB[2] + c_ad * AD[2] {2}.

One may note that AB[1] * {2} - AB[2] * {1} implies that

AB[1] * AO[2] - AB[2] * AO[1] = c_ad * AD[2] * AB[1] - c_ad * AD[1] * AB[2]

that can be further rewritten as

c_ad = (AB[1] * AO[2] - AB[2] * AO[1]) / (AD[2] * AB[1] - AD[1] * AB[2]).

Similarly, AD[1] * {2} - AD[2] * {1} implies that

c_ab = -(AD[1] * AO[2] - AD[2] * AO[1]) / (AD[2] * AB[1] - AD[1] * AB[2]).

This decomposition allows us to determine whether the cell O is inside the rectangle area, since this requires cab and cad to be between 0 and 1.

Note

This function is intended for internal use only.

By convention, the decomposition is done at the top right corner of each cell.

Inputs

ab_ind::Vector{Float64}: Indices representing the edge AB of the rectangle.

ad_ind::Vector{Float64}: Indices representing the edge AD of the rectangle.

a_ind::Vector{Float64}: Indices of the vertex A from which the edges AB and AD start.

area_min_x::Int64: Minimum index in the X direction of the specified area.

area_min_y::Int64: Minimum index in the Y direction of the specified area.

area_length_x::Int64: Number of grid elements in the X direction of the specified area.

area_length_y::Int64: Number of grid elements in the Y direction of the specified area.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

c_ab::Matrix{Float64}: Results of the vector decomposition in terms of the AB component.

c_ad::Matrix{Float64}: Results of the vector decomposition in terms of the AD component.

in_rectangle::Matrix{Bool}: Indicates whether the cell is inside the rectangle area.

n_cell::Int64: Number of cells inside the rectangle area.

Example

ab_ind = [0, 4, 0]
ad_ind = [3, 0, 0]
a_ind = [10, 8, 25]

c_ab, c_ad, in_rectangle, n_cell = _decompose_vector_rectangle(
    ab_ind, ad_ind, a_ind, 15, 12, 8, 7
)

SoilDynamics._decompose_vector_triangle — Method

_decompose_vector_triangle(
    ab_ind::Vector{T}, ac_ind::Vector{T}, a_ind::Vector{T},
    area_min_x::I, area_min_y::I, area_length_x::I, area_length_y::I, tol::T=1e-8
) where {I<:Int64,T<:Float64}

This function performs a vector decomposition on a portion of the horizontal plane where a triangle ABC is located. The position of the triangle is defined by its edges AB and AC , while the specified area extends over [area_min_x, area_min_x + area_length_x] on the X direction and [area_min_y, area_min_y + area_length_y] on the Y direction.

For each cell in the specified area, the function decomposes it into the basis formed by the vectors AB and AC. Let O be the name of a cell, it can then be decomposed as

AO = c_ab * AB + c_ac * AC.

This decomposition leads to a system of 2 equations with 2 unknowns (cab and cac)

AO[1] = c_ab * AB[1] + c_ac * AC[1] {1},
AO[2] = c_ab * AB[2] + c_ac * AC[2] {2}.

One may note that AB[1] * {2} - AB[2] * {1} implies that

AB[1] * AO[2] - AB[2] * AO[1] = c_ac * AC[2] * AB[1] - c_ac * AC[1] * AB[2]

that can be further rewritten as

c_ac = (AB[1] * AO[2] - AB[2] * AO[1]) / (AC[2] * AB[1] - AC[1] * AB[2]).

Similarly, AC[1] * {2} - AC[2] * {1} implies that

c_ab = -(AC[1] * AO[2] - AC[2] * AO[1]) / (AC[2] * AB[1] - AC[1] * AB[2]).

This decomposition allows us to determine whether the cell O is inside the triangle area, since this requires cab and cac to be between 0 and 1, and the sum of cab and cac to be lower than 1.

Note

This function is intended for internal use only.

By convention, the decomposition is done at the top right corner of each cell.

Inputs

ab_ind::Vector{Float64}: Indices representing the edge AB of the triangle.

ac_ind::Vector{Float64}: Indices representing the edge AC of the triangle.

a_ind::Vector{Float64}: Indices of the vertex A from which the edges AB and AC start.

area_min_x::Int64: Minimum index in the X direction of the specified area.

area_min_y::Int64: Minimum index in the Y direction of the specified area.

area_length_x::Int64: Number of grid elements in the X direction of the specified area.

area_length_y::Int64: Number of grid elements in the Y direction of the specified area.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

c_ab::Matrix{Float64}: Results of the vector decomposition in terms of the AB component.

c_ac::Matrix{Float64}: Results of the vector decomposition in terms of the AC component.

in_triangle::Matrix{Bool}: Indicates whether the cell is inside the triangle area.

n_cell::Int64: Number of cells inside the triangle area.

Example

ab_ind = [0, 4, 0]
ac_ind = [3, 0, 0]
a_ind = [10, 8, 25]

c_ab, c_ac, in_triangle, n_cell = _decompose_vector_triangle(
    ab_ind, ac_ind, a_ind, 15, 12, 8, 7
)

SoilDynamics._include_new_body_pos! — Method

_include_new_body_pos!(
    out::SimOut{B,I,T}, ii::I, jj::I, min_h::T, max_h::T, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function updates the bucket position in body at the coordinates (ii, jj). The minimum and maximum heights of the bucket at that position are given by min_h and max_h, respectively. If the given position overlaps with an existing position, then the existing position is updated as the union of the two positions. Otherwise, a new position is added to body.

Note

This function is intended for internal use only.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

ii::Int64: Index of the considered position in the X direction.

jj::Int64: Index of the considered position in the Y direction.

min_h::Float64: Minimum height of the bucket. [m]

max_h::Float64: Maximum height of the bucket. [m]

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_include_new_body_pos!(out, 10, 15, 0.5, 0.6)

SoilDynamics._update_body! — Method

_update_body!(
    area_pos::Vector{Vector{I}}, out::SimOut{B,I,T}, grid::GridParam{I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function updates the bucket position in body following the cells composing area_pos. For each XY position, the first cell found in area_pos corresponds to the minimum height of the bucket, while the last one provides the maximum height. As a result, this function must be called separately for each bucket wall.

Note

This function is intended for internal use only.

area_pos must be sorted and not be empty.

Inputs

area_pos::Vector{Vector{Int64}}: A collection of cell indices specifying where a bucket wall is located.

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)
a = [1.0, 0.0, 0.7]
b = [0.0, 1.0, 0.7]
c = [0.0, 1.0, 0.9]
tri_pos = _calc_triangle_pos(a, b, c, 0.01, grid)

_update_body!(tri_pos, out, grid)