Documentation for intersecting_cells.jl

SoilDynamics._locate_intersecting_cells — Method

_locate_intersecting_cells(
    out::SimOut{B,I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function identifies all the soil cells in the terrain that intersect with the bucket.

Note

This function is intended for internal use only.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

Vector{Vector{Int64}}: Collection of cells indices from the terrain intersecting with the bucket.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_locate_intersecting_cells(out)

SoilDynamics._move_intersecting_body! — Method

_move_intersecting_body!(
    out::SimOut{B,I,T}, grid::GridParam{I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil cells in the terrain that intersect with a bucket. It checks the eight lateral directions surrounding the intersecting soil column and moves the soil to available spaces. If there is insufficient space for all the soil, it incrementally checks the eight directions farther from the intersecting soil column until all the soil has been moved. The process can be illustrated as follows

             ↖   ↑   ↗
               ↖ ↑ ↗
             ← ← O → →
               ↙ ↓ ↘
             ↙   ↓   ↘

Note

This function is intended for internal use only.

The order in which the directions are checked is randomized in order to avoid asymmetrical results.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_move_intersecting_body!(out, grid)

SoilDynamics._move_intersecting_body_soil! — Method

_move_intersecting_body_soil!(
    out::SimOut{B,I,T}, grid::GridParam{I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil cells resting on the bucket that intersect with another bucket layer. It checks the eight lateral directions surrounding the intersecting soil column and moves the soil to available spaces.

It is assumed that the adjacent neighbors always have enough space to accomodate the intersecting soil. This is to avoid the complexity of checking farther cells, which would require considering the potential presence of bucket walls blocking the movement.

In situations where there is insufficient space to accommodate all the intersecting soil, the algorithm currently handles it by allowing the excess soil to simply disappear. This compromise seems to be reasonable as long as the amount of soil disappearing remains negligible.

Note

This function is intended for internal use only.

This function is a work in progress. Additional modifications may be required to improve efficiency and simplicity.

The order in which the directions are checked is randomized in order to avoid asymmetrical results.

By convention, the soil can be moved from the bucket to the terrain even if the bucket is underground.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_move_intersecting_body_soil!(out, grid)

SoilDynamics._move_intersecting_cells! — Method

_move_intersecting_cells!(
    out::SimOut{B,I,T}, grid::GridParam{I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil cells that intersect with the bucket or bucket soil. Since only one bucket is considered, the soil in body_soil is currently not moved. When the simulator supports multiple buckets, this function should account for the possible intersection between the bucket soil and other buckets.

Note

This function is a work in progress.

This function is intended for internal use only.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_move_intersecting_cells!(out, grid)