Documentation for relax.jl

SoilDynamics._locate_unstable_terrain_cell — Method

_locate_unstable_terrain_cell(
    out::SimOut{B,I,T},
    dh_max::T,
    tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function locates all the cells in terrainthat have a height difference larger than dh_max with at least one neighboring cell. Such height difference may indicate that the soil column is unstable. However, it is important to note that this condition is not necessarily indicative of an actual soil instability, as a bucket or the soil resting on it could be supporting the soil column.

Note

This function is intended for internal use only.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

dh_max::Float64: Maximum height difference allowed between two neighboring cells. [m]

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

unstable_cells::Vector{Vector{Int64}}: Collection of cells indices that are possibly unstable.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

unstable_cells = _locate_unstable_terrain_cell(out, 0.1)

SoilDynamics._relax_terrain! — Method

_relax_terrain!(
    out::SimOut{B,I,T},
    grid::GridParam{I,T},
    sim::SimParam{I,T},
    tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil in terrain towards a state closer to equilibrium.

Note

This function is intended for internal use only.

This function is a work in progress.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

sim::SimParam{Int64,Float64}: Struct that stores information related to the simulation.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
sim = SimParam(0.85, 3)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_relax_terrain!(out, grid, sim)