Documentation for utils.jl

This file aggregates a collection of utility functions. Information for some of these functions are provided below.

Checking functions

This corresponds to the functions check_volume and check_soil, which can be used to check the validity of the simulation outputs. It is important to call these functions either before or after the simulation stepping process, as the outputs may be temporarily invalid during the stepping.

The check_volume function verifies mass conservation, ensuring that the total mass of soil is properly conserved throughout the simulation. The check_soil function checks the coherence and integrity of the terrain, bucket soil, and bucket. It ensures that there are no intersections or conflicting cell assignments between these three groups.

Note that these functions are computationally intensive and may slow down the simulation. Therefore, it is recommended to use them primarily for debugging and testing purposes.

Writing functions

This corresponds to the functions write_soil and write_bucket, which can be used to write the outputs into csv files. The write_bucket function generates files starting with bucket, containing the Cartesian coordinates of the four bucket walls' edges. Conversely, the write_soil function generates files starting with terrain for the terrain height and body_soil for the bucket soil height. All the output files are saved in the results folder, with sequentially increasing numbers appended to the filenames.

Note that these functions are primarily designed to be used in the example script provided. Consequently, they have potential limitations that should be considered by the user. Firstly, the file numbers are assigned sequentially. If the write_bucket and write_soil functions are not called together consistently, the file numbers of the bucket and terrain may not correspond. Additionally, this may cause issues if existing files are present in the results folder.

Secondly, the file formats generated by these functions are specifically designed for visualization with Paraview. To visualize the data using a different tool, some change to the file formats may be needed.

_locate_all_non_zeros

The purpose of this function is to provide a list of indices corresponding to the non-zero elements of either the body or body_soil field of the SimOut struct. body and body_soil are a vector of Sparse Matrices indicating the location of the bucket and bucket soil, respectively. Sparse Matrices are used to speed up the code and reduce memory allocation, but this function is still slow and should be used only when necessary.

API

SoilDynamics._init_sparse_array! — Method

_init_sparse_array!(
    sparse_array::Vector{SparseMatrixCSC{T,I}}, grid::GridParam{I,T}
) where {I<:Int64,T<:Float64}

This function reinitializes sparse_array. sparse_array is expected to be either body or body_soil.

Note

This function is intended for internal use only.

Inputs

sparse_array::Vector{SparseMatrixCSC{Float64,Int64}}: Either body or body_soil.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_init_sparse_array!(out.body, grid)

SoilDynamics._locate_all_non_zeros — Method

_locate_all_non_zeros(
    sparse_array::Vector{SparseMatrixCSC{T,I}}
) where {I<:Int64,T<:Float64}

This function returns the indices of all non-zero values in sparse_array. sparse_array is expected to be either body or body_soil.

Note

This function is intended for internal use only.

The first index in the returned vector corresponds to the bucket layer, while the second and third indices are the indices in the X and Y direction, respectively.

Inputs

sparse_array::Vector{SparseMatrixCSC{Float64,Int64}}: Either body or body_soil.

Outputs

Vector{Vector{Int64}}:: Collection of cells indices where sparse_array is non-zero.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

body_soil_pos = _locate_all_non_zeros(out.body_soil)

SoilDynamics._locate_non_zeros — Method

_locate_non_zeros(
    sparse_matrix::SparseMatrixCSC{T,I}
) where {I<:Int64,T<:Float64}

This function returns the indices of all non-zero values in a sparse Matrix.

Note

This function is intended for internal use only.

This implementation is faster than a simple loop.

Inputs

sparse_matrix::SparseMatrixCSC{Float64,Int64}: Input Matrix for which non-zero values should be located.

Outputs

Vector{Vector{Int64}}:: Collection of cells indices where the value of the input Matrix is non-zero.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

non_zeros = _locate_non_zeros(out.body[1])

SoilDynamics._write_vector — Method

_write_vector(
    io, x::T, y::T, z::T
) where {T<:Float64}

This is an utility function to write a 3D vector, such as Cartesian coordinates, into a file.

Note

This function is intended for internal use only.

Inputs

io: Used as a reference to the writing file.

x::Float64: X coordinate of the position to write.

y::Float64: Y coordinate of the position to write.

z::Float64: Z coordinate of the position to write.

Outputs

None

Example

open("file.txt", "a") do io
    write_vector(io, 0.0, 0.0, 1.0)
end

SoilDynamics.calc_normal — Method

calc_normal(
    a::Vector{T}, b::Vector{T}, c::Vector{T}
) where {T<:Float64}

This function calculates the unit normal vector of a plane formed by three points using the right-hand rule.

Note

The input order of the points is important as it determines the sign of the unit normal vector based on the right-hand rule.

Inputs

a::Vector{Float64}: Cartesian coordinates of the first point of the plane. [m]

b::Vector{Float64}: Cartesian coordinates of the second point of the plane. [m]

c::Vector{Float64}: Cartesian coordinates of the third point of the plane. [m]

Outputs

Vector{T}: Unit normal vector of the provided plane. [m]

Example

a = [0.0, 0.0, 0.0]
b = [1.0, 0.5, 0.23]
c = [0.1, 0.2, -0.5]

unit_normal = calc_normal(a, b, c)

SoilDynamics.check_soil — Method

check_soil(
    out::SimOut{B,I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function checks that all the simulation outputs follow the conventions of the simulator. If any inconsistency is found, a warning is issued. The conventions that are checked include:

The terrain should not overlap with the bucket.

The bucket should be properly defined, with its maximum height higher than its minimum height.

The bucket soil should be properly defined, with its maximum height higher than its minimum height.

The two bucket layers should not overlap or touch each other.

One bucket layer should not overlap with all bucket soil layer.

The bucket should not overlap with the corresponding bucket soil layer.

The bucket soil layer should be resting on the corresponding bucket layer.

The bucket should be present when there is bucket soil.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

check_soil(out)

SoilDynamics.check_volume — Method

check_volume(
    out::SimOut{B,I,T}, init_volume::T, grid::GridParam{I,T}
) where {B<:Bool,I<:Int64,T<:Float64}

This function checks that the volume of soil is conserved. The initial volume of soil (init_volume) has to be provided.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

init_volume::Float64: Initial volume of soil in the terrain. [m^3]

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)
init_volume = 0.0

check_volume(out, init_volume, grid)

SoilDynamics.set_RNG_seed! — Method

set_RNG_seed!(
    seed::I=1234
) where {I<:Int64}

This function sets the used RNG seed.

Inputs

seed::Int64: Value of the RNG seed.

Outputs

None

Example

set_RNG_seed!(1234)

SoilDynamics.write_bucket — Method

write_bucket(
    bucket::BucketParam{T}
) where {T<:Float64}

This function writes the position of all bucket faces into a csv located in the "results" directory. The file is named "bucket" followed by the file number.

Inputs

bucket::BucketParam{Float64}: Struct that stores information related to the bucket object.

Outputs

None

Example

o = [0.0, 0.0, 0.0]
j = [0.0, 0.0, 0.0]
b = [0.0, 0.0, -0.5]
t = [1.0, 0.0, -0.5]
bucket = BucketParam(o, j, b, t, 0.5)

write_bucket(bucket)

SoilDynamics.write_soil — Method

write_soil(
    out::SimOut{B,I,T}, grid::GridParam{I,T}
) where {B<:Bool,I<:Int64,T<:Float64}

This function writes the terrain and the bucket soil into a csv located in the "results" directory. terrain and body_soil are saved into files named "terrain" and "body_soil", respectively, followed by the file number.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid

write_soil(out, grid)