Documentation for body_soil.jl

Introductory remarks

The function in this file has two important purposes:

Moving the soil resting on the excavator bucket following the bucket's movement.

Collecting the locations of the bucket soil and storing them in out.body_soil_pos.

While these tasks may initially appear straightforward, they are, in fact, quite complex due to the limited information available and the need to handle edge cases where soil from different locations ends up in the same position.

The first task of the function is to determine the new position of the bucket soil following the bucket's movement. This process involves several steps to accurately calculate the soil's new resting position:

The algorithm starts by taking the current position of the bucket where the soil is resting and calculates the corresponding position in the reference pose of the bucket.

Next, the algorithm applies the updated bucket pose to this reference position to determine the new position where the bucket soil should now be resting.

This process is necessary because the orientation of the bucket is given relative to its reference pose.

The second task is to determine the appropriate bucket layer where the soil should be placed. As it is assumed that the bucket movement between two steps is less than cell_size_xy, the soil is moved to the bucket layer that is within a vertical distance of cell_size_xy from the newly calculated position. However, there can be complications, especially when cell_size_z is significantly lower than cell_size_xy. In such cases, finding the correct bucket layer with the available information may be challenging. If no bucket layer satisfies the above condition, the soil is moved to the terrain.

Note that it is crucial that this function is called after _calc_bucket_pos! and that the bucket pose has not yet been updated in the BucketParam struct.

API

SoilDynamics._update_body_soil! — Method

_update_body_soil!(
    out::SimOut{B,I,T}, pos::Vector{T}, ori::Quaternion{T}, grid::GridParam{I,T},
    bucket::BucketParam{T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil resting on the bucket following its movement. To do so, the movement applied to the base of the soil column is calculated and the soil is moved to this new location. It is however difficult to track accurately each bucket wall. This is currently done by looking at the height difference between the previous and new soil locations, if this height difference is lower than cell_size_xy, it is assumed to be the same bucket wall. Some errors may however be present and further testing is required. If no bucket wall is present, the soil is moved down to the terrain.

The new positions of the soil resting on the bucket are collected into out.body_soil_pos and duplicates are removed.

Note

This function is intended for internal use only.

This function is a work in progress. Some optimization and improvements may be needed.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

pos::Vector{Float64}: Cartesian coordinates of the bucket origin. [m]

ori::Quaternion{Float64}: Orientation of the bucket. [Quaternion]

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

bucket::BucketParam{Float64}: Struct that stores information related to the bucket object.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

pos = [0.5, 0.3, 0.4]
ori = angle_to_quat(0.0, -pi / 2, 0.0, :ZYX)
grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
o = [0.0, 0.0, 0.0]
j = [0.0, 0.0, 0.0]
b = [0.0, 0.0, -0.5]
t = [1.0, 0.0, -0.5]
bucket = BucketParam(o, j, b, t, 0.5)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_update_body_soil!(out, pos, ori, grid, bucket)