Documentation for intersecting_cells.jl

SoilDynamics._locate_intersecting_cells — Method

_locate_intersecting_cells(
    out::SimOut{B,I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function identifies all the soil cells in the terrain that intersect with the bucket.

Note

This function is intended for internal use only.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

Vector{Vector{Int64}}: Collection of cells indices from the terrain intersecting with the bucket.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_locate_intersecting_cells(out)

SoilDynamics._move_body_soil! — Method

_move_body_soil!(
    out::SimOut{B,I,T}, ind_p::I, ii_p::I, jj_p::I, max_h::T, ii_n::I, jj_n::I,
    h_soil::T, wall_presence::B, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function tries to move the soil cells resting on the bucket layer ind_p at the location (ii_p, jj_p) to a new location at (ii_n, jj_n).

This function can be separated into three main scenarios:

If all the soil can be moved to the new location (either on the terrain or on the bucket), the soil is moved and the value of h_soil is set to zero.

If a bucket wall is blocking the movement, the wall_presence parameter is set to true.

If there is insufficient space to move all the soil but no bucket wall is blocking the movement, the function updates the values for the new location and adjusts h_soil accordingly.

This function is designed to be used iteratively by _move_intersecting_body_soil! until all intersecting soil cells are moved.

Note

This function is intended for internal use only.

By convention, the soil can be moved from the bucket to the terrain even if the bucket is underground.

In cases where the soil should be moved to the terrain, all soil is moved regardless of the available space. If this movement induces intersecting soil cells, it will be resolved by the _move_intersecting_body! function.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

ind_p::Int64: Index of the previous considered bucket layer.

ii_p::Int64: Index of the previous considered position in the X direction.

jj_p::Int64: Index of the previous considered position in the Y direction.

max_h::Float64: Maximum height authorized for the movement. [m]

ii_n::Int64: Index of the new considered position in the X direction.

jj_n::Int64: Index of the new considered position in the Y direction.

h_soil::Float64: Height of the soil column left to be moved. [m]

wall_presence::Bool: Indicates whether a bucket wall is blocking the movement.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

ind_p::Int64: Index of the new considered bucket layer.

ii_p::Int64: Index of the new considered position in the X direction.

jj_p::Int64: Index of the new considered position in the Y direction.

h_soil::Float64: Height of the soil column left to be moved. [m]

wall_presence::Bool: Indicates whether a bucket wall is blocking the movement.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

 ind_p, ii_p, jj_p, h_soil, wall_presence = _move_body_soil!(
     out, 1, 10, 15, 0.2, 10, 16, 0.2, true
 )

SoilDynamics._move_intersecting_body! — Method

_move_intersecting_body!(
    out::SimOut{B,I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil cells in the terrain that intersect with a bucket. It checks the eight lateral directions surrounding the intersecting soil column and moves the soil to available spaces. If there is insufficient space for all the soil, it incrementally checks the eight directions farther from the intersecting soil column until all the soil has been moved. The process can be illustrated as follows

             ↖   ↑   ↗
               ↖ ↑ ↗
             ← ← O → →
               ↙ ↓ ↘
             ↙   ↓   ↘

Note

This function is intended for internal use only.

The order in which the directions are checked is randomized in order to avoid asymmetrical results.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_move_intersecting_body!(out)

SoilDynamics._move_intersecting_body_soil! — Method

_move_intersecting_body_soil!(
    out::SimOut{B,I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil cells resting on the bucket that intersect with another bucket layer. It checks the eight lateral directions surrounding the intersecting soil column and moves the soil to available spaces.

The algorithm follows an incremental approach, checking directions farther from the intersecting soil column until it reaches a bucket wall blocking the movement or until all the soil has been moved. If the movement is blocked by a bucket wall, the algorithm explores another direction.

In cases where the soil should be moved to the terrain, all soil is moved regardless of the available space. If this movement induces intersecting soil cells, it will be resolved by the _move_intersecting_body! function.

In rare situations where there is insufficient space to accommodate all the intersecting soil, the algorithm currently handles it by allowing the excess soil to simply disappear. This compromise seems to be reasonable as long as the amount of soil disappearing remains negligible.

Note

This function is intended for internal use only.

The order in which the directions are checked is randomized in order to avoid asymmetrical results.

By convention, the soil can be moved from the bucket to the terrain even if the bucket is underground.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_move_intersecting_body_soil!(out)

SoilDynamics._move_intersecting_cells! — Method

_move_intersecting_cells!(
    out::SimOut{B,I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves all soil cells in terrain and in body_soil that intersect with the bucket or with another soil cell.

Note

This function is intended for internal use only.

_move_intersecting_body_soil! must be called before _move_intersecting_body!, otherwise some intersecting soil cells may remain.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_move_intersecting_cells!(out)