Documentation for relax.jl

SoilDynamics._check_unstable_body_cell — Method

_check_unstable_body_cell(
    out::SimOut{B,I,T}, ii::I, jj::I, ind::I, ii_c::I, jj_c::I, h_min::T, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function checks the stability of a soil column in the soil layer ind of body_soil at (ii, jj) compared to one of its neighbor at (ii_c, jj_c). In case of instability, the function returns a two-digit number (status) that provides information on how the soil should avalanche. The interpretation of the two-digit number is described below.

The first digit indicates the potential presence of the bucket:

1 when the first bucket layer is present.

2 when the second bucket layer is present.

3 when the two bucket layers are present.

4 when no bucket layer is present.

The second digit indicates the layer where the soil should avalanche:

0 when it is the terrain (no bucket is present).

1 when it is the second bucket soil layer.

2 when it is the second bucket layer.

3 when it is the first bucket soil layer.

4 when it is the first bucket layer.

The combination of these two digits provides a comprehensive description of how the soil should avalanche in different scenarios.

Note

This function is intended for internal use only.

Not all combinations for status are possible. Some combinations, such as 41 and 23 are impossible.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

ii::Int64: Index of the considered cell in the X direction.

jj::Int64: Index of the considered cell in the Y direction.

ind::Int64: Index of the considered soil layer.

ii_c::Int64: Index of the neighboring cell in the X direction.

jj_c::Int64: Index of the neighboring cell in the Y direction.

h_min::Float64: Minimum allowed height for a stable configuration. [m]

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

status::Int64: Two-digit number indicating how the soil should avalanche. 0 is returned if the soil column is stable.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

status = _check_unstable_body_cell(out, 10, 14, 1, 10, 15, -0.1)

SoilDynamics._check_unstable_terrain_cell — Method

_check_unstable_terrain_cell(
    out::SimOut{B,I,T}, ii_c::I, jj_c::I, h_min::T, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function checks the stability of a soil column in terrain compared to one of its neighbor (ii_c, jj_c). In case of instability, the function returns a three-digit number (status) that provides information on how the soil should avalanche. The interpretation of the three-digit number is described below.

The first digit indicates the potential presence of the bucket:

1 when the first bucket layer is present.

2 when the second bucket layer is present.

3 when the two bucket layers are present.

4 when no bucket layer is present.

The second digit indicates the layer at the top where the soil should avalanche:

0 when it is the terrain (no bucket is present).

1 when it is the second bucket soil layer.

2 when it is the second bucket layer.

3 when it is the first bucket soil layer.

4 when it is the first bucket layer.

The third digit indicates whether the soil should avalanche below or above the bucket:

0 when there is no bucket.

1 when the soil should avalanche below the bucket.

2 when the soil should avalanche on the top of the bucket.

The combination of these three digits provides a comprehensive description of how the soil should avalanche in different scenarios.

Note

This function is intended for internal use only.

Not all combinations for status are possible. Some combinations, such as 401, 231 and 220, are impossible.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

ii_c::Int64: Index of the neighboring cell in the X direction.

jj_c::Int64: Index of the neighboring cell in the Y direction.

h_min::Float64: Minimum allowed height for a stable configuration. [m]

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

status::Int64: Three-digit number indicating how the soil should avalanche. 0 is returned if the soil column is stable.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

status = _check_unstable_terrain_cell(out, 10, 15, -0.1)

SoilDynamics._locate_unstable_terrain_cell — Method

_locate_unstable_terrain_cell(
    out::SimOut{B,I,T}, dh_max::T, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function locates all the cells in terrainthat have a height difference larger than dh_max with at least one neighboring cell. Such height difference may indicate that the soil column is unstable. However, it is important to note that this condition is not necessarily indicative of an actual soil instability, as a bucket or the soil resting on it could be supporting the soil column.

Note

This function is intended for internal use only.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

dh_max::Float64: Maximum height difference allowed between two neighboring cells. [m]

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

unstable_cells::Vector{Vector{Int64}}: Collection of cells indices that are possibly unstable.

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

unstable_cells = _locate_unstable_terrain_cell(out, 0.1)

SoilDynamics._relax_body_soil! — Method

_relax_body_soil!(
    out::SimOut{B,I,T}, grid::GridParam{I,T}, sim::SimParam{I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil in body_soil towards a state closer to equilibrium. The soil stability is determined by the repose_angle. If the slope formed by two neighboring soil columns exceeds the repose_angle, it is considered unstable, and the soil from the higher column should avalanche to the neighboring column to reach an equilibrium state.

By convention, this function only checks the stability of the soil in the four adjacent cells: ↑ ← O → ↓

The diagonal directions are not checked for simplicity and performance reasons.

This function only moves the soil when the following conditions are met:

(1) The soil column in the neighboring cell is low enough. (2) There is space on the top of the neighboring soil column.

Note

This function is intended for internal use only.

This function is a work in progress.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

sim::SimParam{Int64,Float64}: Struct that stores information related to the simulation.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
sim = SimParam(0.85, 3, 4)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_relax_body_soil!(out, grid, sim)

SoilDynamics._relax_terrain! — Method

_relax_terrain!(
    out::SimOut{B,I,T}, grid::GridParam{I,T}, sim::SimParam{I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil in terrain towards a state closer to equilibrium. The soil stability is determined by the repose_angle. If the slope formed by two neighboring soil columns exceeds the repose_angle, it is considered unstable, and the soil from the higher column should avalanche to the neighboring column to reach an equilibrium state.

By convention, this function only checks the stability of the soil in the four adjacent cells: ↑ ← O → ↓

The diagonal directions are not checked for simplicity and performance reasons.

This function only moves the soil when the following conditions are met:

(1) The soil column in the neighboring cell is low enough. (2) Either: (a) The bucket is not on the soil, meaning there is space between the terrain and the bucket, or there is no bucket. (b) The bucket is on the terrain, but the combination of the bucket and bucket soil is not high enough to prevent soil avalanche.

In case (2a), the soil will avalanche on the terrain, while in case (2b), the soil will avalanche on the bucket.

Note

This function is intended for internal use only.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

sim::SimParam{Int64,Float64}: Struct that stores information related to the simulation.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
sim = SimParam(0.85, 3, 4)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_relax_terrain!(out, grid, sim)

SoilDynamics._relax_unstable_body_cell! — Method

_relax_unstable_body_cell!(
    out::SimOut{B,I,T}, status::I, new_body_soil_pos::Vector{Vector{I}}, dh_max::T,
    ii::I, jj::I, ind::I, ii_c::I, jj_c::I, grid::GridParam{I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil from the soil layer ind of body_soil at (ii, jj) to the soil column in (ii_c, jj_c). The precise movement depends on the status number as explained in the _check_unstable_body_cell function.

The soil is moved such that the slope formed by the two neighboring soil columns is equal to the repose_angle, provided that the bucket is not preventing this configuration.

Note

This function is intended for internal use only.

It is assumed that the given status is accurate, so no extra checks are present.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

status::Int64: Three-digit number indicating how the soil should avalanche.

new_body_soil_pos::Vector{Vector{Int64}}: Queue to append new bodysoilpos.

dh_max::Float64: Maximum height difference allowed between two neighboring cells. [m]

ii::Int64: Index of the considered cell in the X direction.

jj::Int64: Index of the considered cell in the Y direction.

ind::Int64: Index of the considered soil layer.

ii_c::Int64: Index of the neighboring cell in the X direction.

jj_c::Int64: Index of the neighboring cell in the Y direction.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

new_body_soil_pos = Vector{Vector{Int64}}()
grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_relax_unstable_body_cell!(out, 40, new_body_soil_pos, 0.1, 10, 15, 1, 10, 14, grid)

SoilDynamics._relax_unstable_terrain_cell! — Method

_relax_unstable_terrain_cell!(
    out::SimOut{B,I,T}, status::I, dh_max::T, ii::I, jj::I, ii_c::I, jj_c::I,
    grid::GridParam{I,T}, tol::T=1e-8
) where {B<:Bool,I<:Int64,T<:Float64}

This function moves the soil from the terrain at (ii, jj) to the soil column in (ii_c, jj_c). The precise movement depends on the status number as explained in the _check_unstable_terrain_cell function.

The soil is moved such that the slope formed by the two neighboring soil columns is equal to the repose_angle. When the bucket is preventing this configuration, the soil avalanche below the bucket to fill the space under it.

Note

This function is intended for internal use only.

It is assumed that the given status is accurate, so no extra checks are present.

Inputs

out::SimOut{Bool,Int64,Float64}: Struct that stores simulation outputs.

status::Int64: Three-digit number indicating how the soil should avalanche.

dh_max::Float64: Maximum height difference allowed between two neighboring cells. [m]

ii::Int64: Index of the considered cell in the X direction.

jj::Int64: Index of the considered cell in the Y direction.

ii_c::Int64: Index of the neighboring cell in the X direction.

jj_c::Int64: Index of the neighboring cell in the Y direction.

grid::GridParam{Int64,Float64}: Struct that stores information related to the simulation grid.

tol::Float64: Small number used to handle numerical approximation errors.

Outputs

None

Example

grid = GridParam(4.0, 4.0, 3.0, 0.05, 0.01)
terrain = zeros(2 * grid.half_length_x + 1, 2 * grid.half_length_y + 1)
out = SimOut(terrain, grid)

_relax_unstable_terrain_cell!(out, 131, 0.1, 10, 15, 10, 14, grid)